requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.7.0
ariadne==0.23.0
asyncpg==0.27.0
click==8.1.7
fastapi==0.103.0
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==3.2.5
graphql-relay==2.0.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
numpy==1.26.4
packaging==24.2
pandas==2.1.3
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
Rx==1.6.3
singledispatch==3.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.27.0
strawberry-graphql==0.253.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.22.0
